changed Ok and Err to Reply class. initial attempt of the stat command
switching to a reply class allows us to build a reply object which may
or may not have a message or parameters attached to it. like we need for
STAT
+OK 2 306 which returns OK and the number of messages, etc.

Currently the tests are failing because the Regex we use in the
Replies.replyFromString method doesnt like the plus-sign in the "+OK"
reply.